数据库表时间戳和非枚举类型的Id全部从Integer更改为Long，报名表增加活动的外键关联。

+++ engine-dao/src/main/java/com/graduate/engine/model/ActMilestone.java
@@ -4,17 +4,17 @@ import lombok.Data;
 
 @Data
 public class ActMilestone {
-    private Integer milestoneId;
+    private Long milestoneId;
 
-    private Integer actId;
+    private Long actId;
 
     private String milestoneName;
 
     private String milestoneMemo;
 
-    private Integer milestoneFrom;
+    private Long milestoneFrom;
 
-    private Integer milestoneTo;
+    private Long milestoneTo;
 
     private String milestoneReport;
