+++ engine-dao/src/main/java/com/graduate/engine/model/Institute.java
@@ -24,7 +24,7 @@ public class Institute {
 
     private String instOthers;
 
-    private Integer instRegisterDate;
+    private Long instRegisterDate;
 
     private String instIntroduction;
